moved the constructHtml function (for server-side functions) into the utils folder
